refactor(debug_server): extract K144 endpoint family to debug_server_m5.{c,h} (refs #332, Wave 23b)
Audit finding A2 (P1) — `main/debug_server.c` is 4,520 LOC across 72
endpoints, with K144 / camera / video-call / settings / voice / OTA
families all in one file.  Adding any new endpoint touches the same
file every other PR touches; compile-time coupling forces a full
debug-server rebuild for any debug-surface tweak.

This PR establishes the per-family split pattern by extracting the
**K144 endpoint family** as the proof of concept:

  GET  /m5          — diagnostic snapshot
  POST /m5/refresh  — force sys.hwinfo refresh
  GET  /m5/models   — sys.lsmode model registry
  POST /m5/reset    — voice_onboard_reset_failover()

Three new files:

  main/debug_server_internal.h   — shared `tab5_debug_check_auth()`
                                    declaration; will be reused by
                                    every per-family extract that
                                    follows.
  main/debug_server_m5.h         — `void debug_server_m5_register(httpd_handle_t)`
  main/debug_server_m5.c         — 4 handlers + 2 cache-refresh helpers +
                                    5 statics, all moved verbatim from
                                    debug_server.c.  ~283 LOC.

debug_server.c changes:
- Drops 232 LOC (the K144 block + 4 inline URI definitions + 4
  register calls).  Net 4,520 → 4,293 LOC.
- Drops `voice_m5_llm.h` + `voice_onboard.h` includes (only K144 used
  these — the rest of debug_server.c never touched the M5 API).
- Adds `debug_server_internal.h` + `debug_server_m5.h` includes.
- Adds a 3-line `tab5_debug_check_auth()` wrapper around the existing
  static `check_auth()` so the 65 historic call sites stay untouched
  (extract before decompose: this PR moves K144; renaming check_auth
  across 65 sites is a *separate* concern that doesn't ship here).
- Replaces the 4 K144 URI definitions in `tab5_debug_server_start()`
  with a single `debug_server_m5_register(server)` call.

CMakeLists.txt: + `debug_server_m5.c` to main's source list.

## Why this is small + safe

Mechanical move, identical behavior — handlers + helpers + statics
moved verbatim with no internal restructuring (CLAUDE.md "extract
before decompose").  The only call-site change is `check_auth(req)` →
`tab5_debug_check_auth(req)` inside the moved file.

## Pattern for the remaining 6 families (subsequent PRs)

The other endpoint clusters (camera + screenshot, video + call,
voice + chat + mode, settings + nvs, OTA, codec) follow the same
shape:

  1. New `debug_server_<family>.{c,h}`
  2. Move handlers + family-local statics over verbatim
  3. Replace inline `check_auth` calls with `tab5_debug_check_auth`
  4. New `debug_server_<family>_register(httpd_handle_t)` function
  5. Replace inline URI defs + register calls in
     `tab5_debug_server_start` with a single family register call
  6. CMakeLists += new .c file

Each subsequent extract is a self-contained ~150-300 LOC PR.

## Verification on hardware

```
build:  clean — only the 3 new files compiled, debug_server.c relinked.
flash:  OK to /dev/ttyACM0, boot reached ~17s.
GET  /m5           → {chain_active:false, failover_state:1 (probing),
                       hwinfo:{valid:false}, version:"", uart_baud:115200}
POST /m5/refresh   → same shape (refresh path delegates to status)
GET  /m5/models    → {valid:false, count:0}  (K144 not stacked on this
                                              dev unit; route resolves)
POST /m5/reset     → {status:"queued", failover_state:2}

story_smoke (14 steps):  14/14 PASS in 168s
                         (Local mode 'what time is it?' round-trip,
                          camera capture, settings nav — full e2e
                          path exercises check_auth + every other
                          endpoint family that wasn't touched here)
```

The 4 K144 endpoints respond with the exact same JSON shape as
pre-extraction; behavior is identical.
